requirements.txt
agno
openai
python-dotenv
ipython
